+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [1.0.12] - TBD
 
+- Add --example-status-persistence option to mimic rspec example status persistence runner config and to support --only-failures on repeat runs
+
 ## [1.0.11] - 2026-08-03
 
 - Fix a bug in 2.x that would cause an exception because of the 3.x syntax usage (`_1`) in the formatter base